Move current bcode functions to v2 namespace
