+++ src/custom-functions/v2/bcode-indicator.ts
